requirements.txt
flask==3.0.3
numpy
joblib==1.4.2
pandas
yfinance
python-dotenv
requests
tensorflow==2.12.0
keras==3.4.1
